Prevent race condition in draining the pool.
For the old solution to work, the while condition
had to be executed atomically, which is hard to
enforce.

Now, instead, keep better track of the counter:
increase the counter in the *old* thread
and decrease it in the *new*.

That way, when the count is finally zero,
it is certain that all threads are thru.