Avoid re-downloading repositories.
Also: remove opt parse dependency, it isn't used.
Also: refactor the writing of du_out, it was
duplicate in both branches of an if.
Also: if an svn repo exists already -- for now,
just skip, as it saves a lot of time.

Also, ensure that REPO_PATH exists.
That is necessary to use the tool standalone.